+++ components/AppFooter.vue
@@ -4,9 +4,9 @@
   <footer class="font-inter bg-[#101113] py-4 text-center text-sm border-t border-gray-700 text-white">
     <div class="flex flex-col gap-3 text-gray-400">
       <p class="text-lg font-bold">RIZQYA <br> ADZRA.</p>
-      <p>&copy; 2025 All rights reserved. Made with love featuring NuxtJs</p>
+      <p>{{ $t('footer_1') }}</p>
       <div>
-        <p>For inquiries, collaboration, and job offers please contact me through e-mail:</p>
+        <p>{{ $t('footer_2') }}</p>
         <p class="font-bold">rizqyaadzra68@gmail.com</p>
       </div>
     </div>