+++ components/VideoPlay.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="relative w-full overflow-hidden bg-black">
     <!-- Wrapper untuk ratio 21:9 -->
-    <div class="relative w-full" style="aspect-ratio: 21 / 9;">
+    <div class="relative w-full" style="aspect-ratio: 23 / 9;">
       <video
         autoplay
         muted
@@ -19,8 +19,8 @@
 
       <!-- Text -->
       <div class="absolute inset-0 flex flex-col gap-3 items-center justify-center z-20 text-white -translate-y-11">
-        <p class="font-jaro text-8xl font-bold">PORTFOLIO.</p>
-        <p class="font-inter font-bold text-xs">NAME'S RIZQYA (JUST CALL ME KIYAA)</p>
+        <p class="font-jaro text-8xl font-bold">{{ $t('welcome') }}</p>
+        <p class="font-inter font-bold text-xs">{{ $t('introduction') }}</p>
       </div>
     </div>
   </div>