+++ layouts/default.vue
@@ -1,13 +1,14 @@
 <template>
   <div class="flex flex-col min-h-screen dark:bg-black">
+    
+    <div class="sticky top-0 z-50">
+      <AppHeader />
+    </div>
 
-    <AppHeader />
-
-    <main class="overflow-hidden">
+    <main class="flex-1 overflow-hidden mb-24">
       <slot />
     </main>
-    
-    <AppFooter />
 
+    <AppFooter />
   </div>
 </template>