+++ pages/certificate.vue
@@ -1,13 +1,12 @@
 <template>
   <div>
-    
+    <p class="font-bold text-blue-600">
+      apa kabar
+
+    </p>
   </div>
 </template>
 
 <script lang="ts" setup>
 
 </script>
-
-<style>
-
-</style>