+++ CHANGELOG.md
@@ -9,4 +9,11 @@
 - Added google/fonts for Inter, Jaro, and Bebas Neue
 - Fixed typo in nuxt.config.js
 
+### Feat V1.0.3
+- Added Language Switch between Indonesia, English, and Deutsch 
+- Added About Me Subpage
+- Added Abilities Subpage
+- Added Tooltip Component
+
+